backend/requirements.txt
# ============================================
# FRAMEWORK
# ============================================
fastapi==0.109.0
uvicorn[standard]==0.27.0
python-multipart==0.0.6
pydantic==2.5.3
pydantic-settings==2.1.0

# ============================================
# LLM & LANGCHAIN
# ============================================
langchain==0.1.4
langchain-community==0.0.16
langchain-openai==0.0.5
openai==1.10.0

# ============================================
# VECTOR STORE
# ============================================
qdrant-client>=1.10.0

# ============================================
# EMBEDDINGS
# ============================================
sentence-transformers==2.3.1
torch==2.1.2
transformers==4.36.2

# ============================================
# DATABASE
# ============================================
sqlalchemy==2.0.25
psycopg2-binary==2.9.9
alembic==1.13.1
asyncpg==0.29.0

# ============================================
# REDIS (CACHE)
# ============================================
redis==5.0.1
hiredis==2.3.2

# ============================================
# OCR
# ============================================
paddleocr==2.7.0.3
paddlepaddle-gpu==2.6.0
pdf2image==1.17.0
pillow==10.2.0
pytesseract==0.3.10
opencv-python-headless==4.10.0.84

# ============================================
# PDF PROCESSING 
# ============================================
PyMuPDF==1.20.2
pdfplumber==0.10.3
pypdf2==3.0.1
python-docx==1.1.2

# ============================================
# PARALLELIZATION
# ============================================
ray[default]==2.9.1

# ============================================
# SHAREPOINT / MICROSOFT GRAPH
# ============================================
msal==1.26.0
requests==2.31.0

# ============================================
# WEB SCRAPING
# ============================================
playwright==1.41.0
trafilatura==1.6.3
beautifulsoup4==4.12.3
readability-lxml==0.8.1
lxml_html_clean

# ============================================
# MONITORING & LOGGING
# ============================================
prometheus-client==0.19.0
python-json-logger==2.0.7
slowapi==0.1.9

# ============================================
# UTILITIES
# ============================================
python-dotenv==1.0.0
tqdm==4.66.1
httpx==0.26.0

# ============================================
# TESTING (OPCIONAL)
# ============================================
pytest==7.4.4
pytest-asyncio==0.23.3
pytest-cov==4.1.0

# ============================================
# HYBRID SEARCH & RERANKING
# ============================================
scikit-learn==1.4.0
numpy>=1.24.0

# ============================================
# WEB SEARCH
# ============================================
duckduckgo-search>=6.0.0
